Check for forbidden overrides
